Refactor connector base class
